improved gyroscope
The code now tries to wait until the robot is complete with its spin,
as the robot remains to spin a litte even if the xbox controller does
not tell it to. This means that the robot will now (most of the time)
start heading straight in the direction that it's facing after a turn,
while before it moved towards the heading it was at when the controller
buttons were released.