Update gyro-v2.1 before merge onto master
Every feature except checking the rotational speed before resetting the
gyro has been copied from the master branch. Modifications:
- ballCannon speed is no longer variable
- left out some debugging messages because they don't seem necessary
anymore
- SmartDashboard will now print the current being drawn by channel 13,
the ball cannon
- IMPORTANT: switched the ports of the pwm channels back to the way they
originally were, left on 0 and right on 1. The PHYSICAL configuration of
the robot has been changed to match this. I think that doing this
will avoid confusion in the future.