Added MySafety and MyGamepad.
Major overhaul. MySafety is a new class that XboxDrive, RopeClimber, and
Cannon subclass. It is my attempt to provide a more systematic way to
disable motors when higher-priority system are powered on. To stop a
motor, call the system's setSafe(false) method. This will set an
instance variable that set(value) checks before outputting to motors.
MyGamepad is a subclass of XboxController, with added functionality to
check the toggle state of the A button. Also, Cannon and RopeClimber are
no longer aware of the gamepad. Robot checks the gamepad and calls their
methods. XboxDrive is still aware of the gamepad, though.