OMG Overhaul.
Removed MyGamepad. Changed MyHID from an interface to a class that allows me to use any input device I want without modifying Robot.java. It contains an enum called Type, which is used to map the correct raw buttons and raw axes to the correct functions for each unique controller. Thanks to this, the functions for getting input can be named after what they do on the robot. Nice! Rewrote Robot.javaand XboxDrive.java with these new functions in mind. I still need to map the conroller buttons an axes for each controller.